Register components for Dwrf, Parquet, FileSystem during HiveConnector registration
