+++ velox/connectors/hive/CMakeLists.txt
@@ -26,11 +26,14 @@ add_library(
 target_link_libraries(
   velox_hive_connector
   velox_connector
+  velox_dwio_catalog_fbhive
   velox_dwio_dwrf_reader
   velox_dwio_dwrf_writer
+  velox_dwio_parquet_reader
+  velox_dwio_parquet_writer
   velox_file
   velox_hive_partition_function
-  velox_dwio_catalog_fbhive)
+  velox_s3fs)
 
 add_library(velox_hive_partition_function HivePartitionFunction.cpp)
