+++ velox/connectors/hive/HiveConnector.cpp
@@ -18,6 +18,17 @@
 
 #include "velox/common/base/Fs.h"
 #include "velox/connectors/hive/HivePartitionFunction.h"
+// Meta's buck build system needs this check.
+#ifdef VELOX_ENABLE_S3
+#include "velox/connectors/hive/storage_adapters/s3fs/RegisterS3FileSystem.h"
+#endif
+#include "velox/dwio/dwrf/reader/DwrfReader.h"
+#include "velox/dwio/dwrf/writer/Writer.h"
+// Meta's buck build system needs this check.
+#ifdef VELOX_ENABLE_PARQUET
+#include "velox/dwio/parquet/RegisterParquetReader.h"
+#include "velox/dwio/parquet/RegisterParquetWriter.h"
+#endif
 #include "velox/expression/FieldReference.h"
 
 #include <boost/lexical_cast.hpp>
@@ -63,6 +74,24 @@ std::unique_ptr<core::PartitionFunction> HivePartitionFunctionSpec::create(
                                  : bucketToPartition_,
       channels_,
       constValues_);
+}
+
+void HiveConnectorFactory::initialize() {
+  static bool once = []() {
+    dwio::common::WriteFileDataSink::registerLocalFileFactory();
+    dwrf::registerDwrfReaderFactory();
+    dwrf::registerDwrfWriterFactory();
+// Meta's buck build system needs this check.
+#ifdef VELOX_ENABLE_PARQUET
+    parquet::registerParquetReaderFactory();
+    parquet::registerParquetWriterFactory();
+#endif
+// Meta's buck build system needs this check.
+#ifdef VELOX_ENABLE_S3
+    filesystems::registerS3FileSystem();
+#endif
+    return true;
+  }();
 }
 
 std::string HivePartitionFunctionSpec::toString() const {
