+++ velox/connectors/hive/storage_adapters/s3fs/tests/S3FileSystemTest.cpp
@@ -15,6 +15,7 @@
  */
 
 #include "connectors/hive/storage_adapters/s3fs/S3FileSystem.h"
+#include "connectors/hive/storage_adapters/s3fs/RegisterS3FileSystem.h"
 #include "connectors/hive/storage_adapters/s3fs/S3Util.h"
 #include "connectors/hive/storage_adapters/s3fs/tests/MinioServer.h"
 #include "velox/common/file/File.h"
@@ -34,7 +35,6 @@ class S3FileSystemTest : public testing::Test {
       minioServer_ = std::make_shared<MinioServer>();
       minioServer_->start();
     }
-    filesystems::registerS3FileSystem();
   }
 
   static void TearDownTestSuite() {
